+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+* `2.0.97` Release - [2.0.97](#2097)
 * `2.0.96` Release - [2.0.96](#2096)
 * `2.0.95` Release - [2.0.95](#2095)
 * `2.0.93` Release - [2.0.93](#2093)
@@ -14,13 +15,17 @@
 * `2.0.78` Release - [2.0.78](#2078)
 * `0.77.x` Releases - [0.77.0](#0770)
 
+## 2.0.97
+
+#### Bug Fixes
+* Log message instead of hierarchy from error if the root error is an `ExposureError` 
+
 ## 2.0.96
 
 #### Features
 * `EMP-12204` Introduced `RBMTheme` to handle common UI elements related to the reference app.
 * `EMP-12206` Added `TableViewDataSource` class to handle `UITableViewDataSource`.
 * `EMP-12206` Added `UITableView+EmptyMessage` to show empty message when `UITableView`  has no data .
-
 
 #### Changes
 * `EMP-12204` Load the project from programmatically created views instead from the storyboard.